requirements.txt
# Discord
discord.py==2.3.2

# LangGraph + LangChain
langgraph==0.2.28
langchain-anthropic==0.2.4
langchain-core==0.3.15

# Scheduler
APScheduler==3.10.4

# Utilities
python-dotenv==1.0.1
aiofiles==23.2.1
gitpython==3.1.43
